Move Proxy-related classes from Data to Proxies; Fix issues where body visibility conflicts with affordance visibility
