+++ Scripts/Proxies/AffordanceTooltip.cs.meta
