+++ Scripts/Proxies/Data/Affordance.cs
