+++ Scripts/Proxies/Data/AffordanceAnimationDefinition.cs
