+++ Scripts/Proxies/Data/AffordanceAnimationDefinition.cs.meta
