+++ Scripts/Proxies/Data/AffordanceDefinition.cs
