+++ Scripts/Proxies/Data/AffordanceDefinition.cs.meta
