+++ Scripts/Proxies/Data/AffordanceVisibilityDefinition.cs
