+++ Scripts/Proxies/Data/AffordanceVisibilityDefinition.cs.meta
