+++ Scripts/Proxies/Data/ProxyAffordanceMap.cs
