+++ Scripts/Proxies/ShakeVelocityTracker.cs
